Only delete aircraft at the runway threshold if they have approach clearance
In other words, if they were told to intercept the localizer but never cleared
to land, they continue flying their present heading past the field...

+++ aircraft.go
@@ -115,7 +115,7 @@ func (ac *Aircraft) Update(wind WindModel, w *World, ep EventPoster) {
 				})
 			}
 		}
-		if passedWaypoint.Delete {
+		if passedWaypoint.Delete && ac.Nav.Approach.Cleared {
 			w.DeleteAircraft(ac, nil)
 		}
 	}
